+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to the "jira-issues" extension will be documented in this fi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.0.0
+## [1.0.0] 2018-11-27
 - Initial release
 
-# 1.0.1 - 1.0.2
-- fix bug where extension doesn't load
+## [1.0.1 - 1.0.3] 2018-12-1
+### Fixed
+- bug where extension doesn't load at all
+- load extension even if configuration is missing with helpful message
+### Changed
+- added more robust error messages throughout